Refactor/kan 18/error handler

See merge request jala-university1/cohort-1/oficial-pt-programa-o-4-apr-221/se-o-b/pes-de-pano!10

+++ src/data/usecases/waypoint/delete-waypoint.js
@@ -1,9 +1,15 @@
+import { NotFoundError, ServerError } from "../../../http/errors.js";
 import { Waypoint } from "../../models/waypoint.js";
 
 export async function deleteWaypoint(id) {
+  const waypoint = await Waypoint.findById(id);
+  if (!waypoint) {
+    throw new NotFoundError(`Waypoint with id ${id} not found`);
+  }
+
   try {
-    return Waypoint.findByIdAndDelete(id);
+    await Waypoint.findByIdAndDelete(id);
   } catch (error) {
-    throw new Error("Error at deleting Waypoint");
+    throw new ServerError("Error at deleting Waypoint");
   }
 }
